refactor(launcher): move detached app spawning into runtime utilities
